+++ NEWS.md
@@ -1,5 +1,12 @@
 # surveydown (development version)
 
+# surveydown 0.11.0
+
+- Questions can now be defined using an external yml file, defaulting to `"questions.yml"`.
+- New `yml` argument in `sd_question()`, defaults to `"questions.yml"`, to allow users which yml file to use for questions. For details, refer to the [Defining Questions](https://surveydown.org/docs/defining-questions) documentation page. 
+- Update to `sd_create_survey()`: now the the `template` argument is by default `"default"`.
+- Update to `sd_create_survey()`: new `template = "questions_yml"` option to create a survey with the `questions_yml` template.
+
 # surveydown 0.10.1
 
 - `sd_create_survey()` now accepts two arguments. The `template` argument is by default `"plain_template"`, which creates a default plain template of surveydown. It also accepts a list of templates that we created in the [templates repo](https://github.com/surveydown-dev/templates/tree/main). The `path` argument is unchanged. It defines the relative path of the template location.